+++ ICD.Connect.Conferencing.Zoom/Controls/ZoomRoomLayoutControl.cs
@@ -56,8 +56,6 @@ namespace ICD.Connect.Conferencing.Zoom.Controls
 
 			Subscribe(m_LayoutComponent);
 			Subscribe(m_CallComponent);
-
-			LayoutAvailable = true;
 		}
 
 		/// <summary>
@@ -130,6 +128,7 @@ namespace ICD.Connect.Conferencing.Zoom.Controls
 			if (m_CallComponent.GetParticipants().Count() <= 1)
 				SelfViewEnabled = true;
 
+			m_LayoutComponent.GetLayoutStatus();
 			m_LayoutComponent.UpdateLayout();
 		}
 
@@ -142,6 +141,7 @@ namespace ICD.Connect.Conferencing.Zoom.Controls
 			layoutComponent.OnSizeChanged += LayoutComponentOnSizeChanged;
 			layoutComponent.OnSelfViewEnabledChanged += LayoutComponentOnSelfViewEnabledChanged;
 			layoutComponent.OnShareThumbChanged += LayoutComponentOnShareThumbChanged;
+			layoutComponent.OnCallLayoutAvailabilityChanged += LayoutComponentOnCallLayoutAvailabilityChanged;
 		}
 
 		private void Unsubscribe(LayoutComponent layoutComponent)
@@ -149,6 +149,7 @@ namespace ICD.Connect.Conferencing.Zoom.Controls
 			layoutComponent.OnSizeChanged -= LayoutComponentOnSizeChanged;
 			layoutComponent.OnSelfViewEnabledChanged -= LayoutComponentOnSelfViewEnabledChanged;
 			layoutComponent.OnShareThumbChanged -= LayoutComponentOnShareThumbChanged;
+			layoutComponent.OnCallLayoutAvailabilityChanged -= LayoutComponentOnCallLayoutAvailabilityChanged;
 		}
 
 		private void LayoutComponentOnShareThumbChanged(object sender, BoolEventArgs eventArgs)
@@ -173,8 +174,16 @@ namespace ICD.Connect.Conferencing.Zoom.Controls
 				m_LayoutComponent.LayoutSize != eZoomLayoutSize.Off;
 		}
 
-		#endregion
+		private void LayoutComponentOnCallLayoutAvailabilityChanged(object sender, GenericEventArgs<ZoomLayoutAvailability> e)
+		{
+			var availability = e.Data;
+			if (availability == null)
+				return;
 
+			LayoutAvailable = availability.IsSupported;
+		}
+
+		#endregion
 		#region Call Component Callbacks
 
 		private void Subscribe(CallComponent callComponent)